requirements.txt
asgiref==3.8.1
Django==4.2.17
django-ckeditor==6.7.2
django-jalali==7.2.0
django-js-asset==2.2.0
django-mysql==4.15.0
jalali_core==1.0.0
jdatetime==5.0.0
mysqlclient==2.2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
